Move printMessageInBox and related utilities to @rushstack/terminal.
