+++ apps/rush-lib/package.json
@@ -53,7 +53,6 @@
     "strict-uri-encode": "~2.0.0",
     "tar": "~5.0.5",
     "true-case-path": "~2.2.1",
-    "wordwrap": "~1.0.0",
     "z-schema": "~3.18.3"
   },
   "devDependencies": {
@@ -77,7 +76,6 @@
     "@types/ssri": "~7.1.0",
     "@types/strict-uri-encode": "2.0.0",
     "@types/tar": "4.0.3",
-    "@types/wordwrap": "1.0.0",
     "@types/z-schema": "3.16.31",
     "jest": "~25.4.0",
     "typescript": "~3.9.7"
