+++ apps/rush-lib/src/utilities/Utilities.ts
@@ -1,13 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
-
 import * as child_process from 'child_process';
 import * as fs from 'fs';
 import * as os from 'os';
-import * as tty from 'tty';
 import * as path from 'path';
-import wordwrap from 'wordwrap';
-import { JsonFile, IPackageJson, FileSystem, FileConstants, Terminal } from '@rushstack/node-core-library';
+import { JsonFile, IPackageJson, FileSystem, FileConstants } from '@rushstack/node-core-library';
 import type * as stream from 'stream';
 import { CommandLineHelper } from '@rushstack/ts-command-line';
 
@@ -308,34 +305,6 @@ export class Utilities {
   }
 
   /**
-   * Returns the width of the console, measured in columns
-   */
-  public static getConsoleWidth(): number {
-    const stdout: tty.WriteStream = process.stdout as tty.WriteStream;
-    if (stdout && stdout.columns) {
-      return stdout.columns;
-    }
-
-    return 80;
-  }
-
-  /**
-   * Applies word wrapping.  If maxLineLength is unspecified, then it defaults to the console
-   * width.
-   */
-  public static wrapWords(text: string, maxLineLength?: number, indent?: number): string {
-    if (!indent) {
-      indent = 0;
-    }
-    if (!maxLineLength) {
-      maxLineLength = Utilities.getConsoleWidth();
-    }
-
-    const wrap: (textToWrap: string) => string = wordwrap(indent, maxLineLength, { mode: 'soft' });
-    return wrap(text);
-  }
-
-  /**
    * Executes the command with the specified command-line parameters, and waits for it to complete.
    * The current directory will be set to the specified workingDirectory.
    */
@@ -623,30 +592,6 @@ export class Utilities {
 
   public static getPackageDepsFilenameForCommand(command: string): string {
     return `package-deps_${command}.json`;
-  }
-
-  public static printMessageInBox(
-    message: string,
-    terminal: Terminal,
-    boxWidth: number = Math.floor(Utilities.getConsoleWidth() / 2)
-  ): void {
-    const maxLineLength: number = boxWidth - 10;
-
-    const wrappedMessage: string = Utilities.wrapWords(message, maxLineLength);
-    const wrappedMessageLines: string[] = wrappedMessage.split('\n');
-
-    // ╔═══════════╗
-    // ║  Message  ║
-    // ╚═══════════╝
-    terminal.writeLine(` ╔${'═'.repeat(boxWidth - 2)}╗ `);
-    for (const line of wrappedMessageLines) {
-      const trimmedLine: string = line.trim();
-      const padding: number = boxWidth - trimmedLine.length - 2;
-      const leftPadding: number = Math.floor(padding / 2);
-      const rightPadding: number = padding - leftPadding;
-      terminal.writeLine(` ║${' '.repeat(leftPadding)}${trimmedLine}${' '.repeat(rightPadding)}║ `);
-    }
-    terminal.writeLine(` ╚${'═'.repeat(boxWidth - 2)}╝ `);
   }
 
   public static async usingAsync<TDisposable extends IDisposable>(
