+++ common/config/rush/repo-state.json
@@ -1,5 +1,5 @@
 // DO NOT MODIFY THIS FILE MANUALLY BUT DO COMMIT IT. It is generated and used by Rush.
 {
-  "pnpmShrinkwrapHash": "342f968435523962819b5505a56ffec0a8b01a45",
+  "pnpmShrinkwrapHash": "5b69259972abbbc86afabdede87a2bc8b644f62a",
   "preferredVersionsHash": "6a96c5550f3ce50aa19e8d1141c6c5d4176953ff"
 }
