+++ common/reviews/api/terminal.api.md
@@ -6,6 +6,7 @@
 
 import { Brand } from '@rushstack/node-core-library';
 import { NewlineKind } from '@rushstack/node-core-library';
+import { Terminal } from '@rushstack/node-core-library';
 
 // @public
 export class CallbackWritable extends TerminalWritable {
@@ -104,6 +105,13 @@ export class NormalizeNewlinesTextRewriter extends TextRewriter {
     readonly newlineKind: NewlineKind;
     // (undocumented)
     process(unknownState: TextRewriterState, text: string): string;
+}
+
+// @public
+export class PrintUtilities {
+    static getConsoleWidth(): number;
+    static printMessageInBox(message: string, terminal: Terminal, boxWidth?: number): void;
+    static wrapWords(text: string, maxLineLength?: number, indent?: number): string;
 }
 
 // @public
