+++ libraries/terminal/src/PrintUtilities.ts
@@ -1,0 +1,68 @@
+// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+// See LICENSE in the project root for license information.
+
+import * as tty from 'tty';
+import wordwrap from 'wordwrap';
+import { Terminal } from '@rushstack/node-core-library';
+
+/**
+ * A collection of utilities for printing messages to the console.
+ *
+ * @public
+ */
+export class PrintUtilities {
+  /**
+   * Returns the width of the console, measured in columns
+   */
+  public static getConsoleWidth(): number {
+    const stdout: tty.WriteStream = process.stdout as tty.WriteStream;
+    if (stdout && stdout.columns) {
+      return stdout.columns;
+    }
+
+    return 80;
+  }
+
+  /**
+   * Applies word wrapping.  If maxLineLength is unspecified, then it defaults to the
+   * console width.
+   */
+  public static wrapWords(text: string, maxLineLength?: number, indent?: number): string {
+    if (!indent) {
+      indent = 0;
+    }
+
+    if (!maxLineLength) {
+      maxLineLength = PrintUtilities.getConsoleWidth();
+    }
+
+    const wrap: (textToWrap: string) => string = wordwrap(indent, maxLineLength, { mode: 'soft' });
+    return wrap(text);
+  }
+
+  /**
+   * Displays a message in the console wrapped in a box UI.
+   */
+  public static printMessageInBox(
+    message: string,
+    terminal: Terminal,
+    boxWidth: number = Math.floor(PrintUtilities.getConsoleWidth() / 2)
+  ): void {
+    const maxLineLength: number = boxWidth - 10;
+    const wrappedMessage: string = PrintUtilities.wrapWords(message, maxLineLength);
+    const wrappedMessageLines: string[] = wrappedMessage.split('\n');
+
+    // ╔═══════════╗
+    // ║  Message  ║
+    // ╚═══════════╝
+    terminal.writeLine(` ╔${'═'.repeat(boxWidth - 2)}╗ `);
+    for (const line of wrappedMessageLines) {
+      const trimmedLine: string = line.trim();
+      const padding: number = boxWidth - trimmedLine.length - 2;
+      const leftPadding: number = Math.floor(padding / 2);
+      const rightPadding: number = padding - leftPadding;
+      terminal.writeLine(` ║${' '.repeat(leftPadding)}${trimmedLine}${' '.repeat(rightPadding)}║ `);
+    }
+    terminal.writeLine(` ╚${'═'.repeat(boxWidth - 2)}╝ `);
+  }
+}
